clippy

+++ sow-core/src/execution/bots.rs
@@ -52,7 +52,7 @@ impl SowEngine {
             for (nx, ny) in neighbors {
                 let owner = self.state.map.owner_id(nx, ny);
                 if owner != player.id {
-                    let is_land = self.state.map.terrain[self.state.map.ref_id(nx as u32, ny as u32)].is_land();
+                    let is_land = self.state.map.terrain[self.state.map.ref_id(nx, ny)].is_land();
                     if !is_land { continue; }
                     targets.push(owner);
                     if owner == 0 {
